initial commit
In this article, we will learn how we can get information about phone numbers like the country name to which the phone number
belongs, or the network service provider name of that phone number just by writing a simple Python program. Python provides a
module named phone numbers for this task. This article is Python port of Google libphonenumber library.